+++ src/main/java/roomescape/controller/AdminController.java
@@ -1,12 +1,8 @@
 package roomescape.controller;
 
-import java.time.LocalTime;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
-import roomescape.model.ReservationTime;
-import roomescape.repository.ReservationTimeRepository;
 
 @Controller
 @RequestMapping("/admin")
