+++ src/main/java/roomescape/service/ReservationService.java
@@ -1,48 +1,44 @@
 package roomescape.service;
 
-import java.time.LocalDate;
-import java.time.LocalTime;
 import java.util.List;
-import java.util.concurrent.atomic.AtomicLong;
 import org.springframework.http.HttpStatus;
-import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
 import roomescape.dto.ReservationRequest;
 import roomescape.dto.ReservationResponse;
 import roomescape.model.Reservation;
 import roomescape.model.ReservationTime;
-import roomescape.repository.ReservationRepository;
-import roomescape.repository.ReservationTimeRepository;
+import roomescape.dao.JdbcReservationDao;
+import roomescape.dao.JdbcReservationTimeDao;
 
 @Service
 public class ReservationService {
 
-    private final ReservationRepository reservationRepository;
-    private final ReservationTimeRepository reservationTimeRepository;
+    private final JdbcReservationDao jdbcReservationDao;
+    private final JdbcReservationTimeDao jdbcReservationTimeDao;
 
-    public ReservationService(ReservationRepository reservationRepository,
-                              ReservationTimeRepository reservationTimeRepository) {
-        this.reservationRepository = reservationRepository;
-        this.reservationTimeRepository = reservationTimeRepository;
+    public ReservationService(JdbcReservationDao jdbcReservationDao,
+                              JdbcReservationTimeDao jdbcReservationTimeDao) {
+        this.jdbcReservationDao = jdbcReservationDao;
+        this.jdbcReservationTimeDao = jdbcReservationTimeDao;
     }
 
     public ReservationResponse addReservation(ReservationRequest reservationRequest) {
-        ReservationTime reservationTime = reservationTimeRepository.findById(reservationRequest.timeId());
+        ReservationTime reservationTime = jdbcReservationTimeDao.findById(reservationRequest.timeId());
         Reservation reservation = reservationRequest.toEntityWithReservationTime(reservationTime);
-        Reservation savedReservation = reservationRepository.save(reservation);
+        Reservation savedReservation = jdbcReservationDao.save(reservation);
         return ReservationResponse.fromEntity(savedReservation);
     }
 
     public void deleteReservation(Long id) {
-        boolean isDeleted = reservationRepository.delete(id);
+        boolean isDeleted = jdbcReservationDao.deleteById(id);
         if (!isDeleted) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, "해당하는 id가 없습니다");
         }
     }
 
     public List<ReservationResponse> getReservations() {
-        return reservationRepository.findAll()
+        return jdbcReservationDao.findAll()
                 .stream()
                 .map(ReservationResponse::fromEntity)
                 .toList();
